+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0
 
 - **_Breaking Change_**: stricter trait bound for `MockLambdaRuntimeApiServer` methods.
+- **_Breaking Change_**: `MockLambdaRuntimeApiServer::passthrough` will create a new client every time.
 
 ## v0.1.1